+++ packetbeat/flows/flows.go
@@ -26,12 +26,6 @@ import (
 	"github.com/elastic/elastic-agent-libs/logp"
 )
 
-type Flows struct {
-	worker     *worker
-	table      *flowMetaTable
-	counterReg *counterReg
-}
-
 // Reporter callback type, to report flow events to.
 type Reporter func([]beat.Event)
 
@@ -42,6 +36,15 @@ const (
 	defaultPeriod  = 10 * time.Second
 )
 
+// Flows holds and publishes network flow information for running processes.
+type Flows struct {
+	worker     *worker
+	table      *flowMetaTable
+	counterReg *counterReg
+}
+
+// NewFlows returns a Flows publishing to pub after enrichment by the given
+// process watcher. Publication timeout and period are specified by config.
 func NewFlows(pub Reporter, watcher procs.ProcessesWatcher, config *config.Flows) (*Flows, error) {
 	duration := func(s string, d time.Duration) (time.Duration, error) {
 		if s == "" {
