+++ packetbeat/flows/worker.go
@@ -21,6 +21,7 @@ import (
 	"encoding/binary"
 	"errors"
 	"net"
+	"sync"
 	"time"
 
 	"github.com/elastic/beats/v7/libbeat/beat"
@@ -28,29 +29,106 @@ import (
 	"github.com/elastic/beats/v7/libbeat/common/flowhash"
 	"github.com/elastic/beats/v7/packetbeat/procs"
 	"github.com/elastic/beats/v7/packetbeat/protos/applayer"
+	"github.com/elastic/elastic-agent-libs/logp"
 	"github.com/elastic/elastic-agent-libs/mapstr"
 )
-
-type flowsProcessor struct {
-	spool    spool
-	watcher  procs.ProcessesWatcher
-	table    *flowMetaTable
-	counters *counterReg
-	timeout  time.Duration
-}
 
 var (
 	ErrInvalidTimeout = errors.New("timeout must be >= 1s")
 	ErrInvalidPeriod  = errors.New("report period must be -1 or >= 1s")
 )
 
-func newFlowsWorker(
-	pub Reporter,
-	watcher procs.ProcessesWatcher,
-	table *flowMetaTable,
-	counters *counterReg,
-	timeout, period time.Duration,
-) (*worker, error) {
+// worker is a generic asynchronous function processor.
+// FIXME: Unexport methods on worker.
+type worker struct {
+	wg   sync.WaitGroup
+	done chan struct{}
+	run  func(*worker)
+}
+
+// newWorker returns a handle to a worker to run fn.
+func newWorker(fn func(w *worker)) *worker {
+	return &worker{
+		done: make(chan struct{}),
+		run:  fn,
+	}
+}
+
+// start starts execution of the worker function.
+func (w *worker) Start() {
+	debugf("start flows worker")
+	w.wg.Add(1)
+	go func() {
+		defer w.finished()
+		w.run(w)
+	}()
+}
+
+// finished decrements the workers working function count. finished
+// must be called the same number of times as start over the lifetime
+// of the worker.
+func (w *worker) finished() {
+	w.wg.Done()
+	logp.Info("flows worker loop stopped")
+}
+
+// stop terminates the function and waits until processing is complete.
+// stop may only be called once.
+func (w *worker) Stop() {
+	debugf("stop flows worker")
+	close(w.done)
+	w.wg.Wait()
+	debugf("stopped flows worker")
+}
+
+// sleep will sleep for the provided duration unless the worker has been
+// stopped. sleep returns whether the worker can continue processing.
+func (w *worker) sleep(d time.Duration) bool {
+	select {
+	case <-w.done:
+		return false
+	case <-time.After(d):
+		return true
+	}
+}
+
+// tick will sleep until the provided ticker fires unless the worker has been
+// stopped. tick returns whether the worker can continue processing.
+func (w *worker) tick(t *time.Ticker) bool {
+	select {
+	case <-w.done:
+		return false
+	case <-t.C:
+		return true
+	}
+}
+
+// periodically will execute fn each tick duration until the worker has been
+// stopped or fn returns a non-nil error.
+func (w *worker) periodically(tick time.Duration, fn func() error) {
+	defer debugf("stop periodic loop")
+
+	ticker := time.NewTicker(tick)
+	defer ticker.Stop()
+	for {
+		cont := w.tick(ticker)
+		if !cont {
+			return
+		}
+
+		err := fn()
+		if err != nil {
+			return
+		}
+	}
+}
+
+// newFlowsWorker returns a worker with a flow lifetime specified by timeout and a
+// reporting intervals specified by period. If period is less than or equal to zero
+// reporting will be done at flow lifetime end.
+// Flows are published via the pub Reporter after being enriched with process information
+// by watcher.
+func newFlowsWorker(pub Reporter, watcher procs.ProcessesWatcher, table *flowMetaTable, counters *counterReg, timeout, period time.Duration) (*worker, error) {
 	oneSecond := 1 * time.Second
 
 	if timeout < oneSecond {
@@ -96,17 +174,73 @@ func newFlowsWorker(
 	return makeWorker(processor, tickDuration, ticksTimeout, ticksPeriod, 10)
 }
 
-func makeWorker(
-	processor *flowsProcessor,
-	tickDuration time.Duration,
-	ticksTimeout, ticksPeriod int,
-	align int64,
-) (*worker, error) {
+// gcd returns the greatest common divisor of a and b.
+func gcd(a, b int64) int64 {
+	// FIXME: This can be rewritten as:
+	//
+	//   func gcd(a, b time.Duration) time.Duration {
+	//   	for b != 0 {
+	//   		a, b = b, a%b
+	//   	}
+	//   	return a
+	//   }
+	//
+	// with improved performance for all cases except a == b where costs are <10ns/op.
+	// This function is only called once on flow worker creation.
+	if a < 0 || b < 0 {
+		return 0
+	}
+
+	switch {
+	case a == b:
+		return a
+	case a == 0:
+		return b
+	case b == 0:
+		return a
+	}
+
+	shift := uint(0)
+	for (a&1) == 0 && (b&1) == 0 {
+		shift++
+		a /= 2
+		b /= 2
+	}
+
+	for (a & 1) == 0 {
+		a = a / 2
+	}
+
+	// a is always odd
+	for {
+		for (b & 1) == 0 {
+			b = b / 2
+		}
+
+		// both a and b are odd. guaranteed b >= a
+		if a > b {
+			a, b = b, a
+		}
+		b -= a
+
+		if b == 0 {
+			break
+		}
+	}
+
+	// restore common factors of 2
+	return a << shift
+}
+
+// makeWorker returns a worker that runs processor.execute each tick. Each timeout'th tick,
+// the worker will check flow timeouts and each period'th tick, the worker will report flow
+// events to be published.
+func makeWorker(processor *flowsProcessor, tickDuration time.Duration, ticksTimeout, ticksPeriod int, align int64) (*worker, error) {
 	return newWorker(func(w *worker) {
 		defer processor.execute(w, false, true, true)
 
 		if align > 0 {
-			// round time to nearest 10 seconds for alignment
+			// Wait until the current time rounded up to nearest align seconds.
 			aligned := time.Unix(((time.Now().Unix()+(align-1))/align)*align, 0)
 			waitStart := time.Until(aligned)
 			debugf("worker wait start(%v): %v", aligned, waitStart)
@@ -137,6 +271,14 @@ func makeWorker(
 			return nil
 		})
 	}), nil
+}
+
+type flowsProcessor struct {
+	spool    spool
+	watcher  procs.ProcessesWatcher
+	table    *flowMetaTable
+	counters *counterReg
+	timeout  time.Duration
 }
 
 func (fw *flowsProcessor) execute(w *worker, checkTimeout, handleReports, lastReport bool) {
@@ -191,25 +333,14 @@ func (fw *flowsProcessor) execute(w *worker, checkTimeout, handleReports, lastRe
 	fw.spool.flush()
 }
 
-func (fw *flowsProcessor) report(
-	w *worker,
-	ts time.Time,
-	flow *biFlow,
-	isOver bool,
-	intNames, uintNames, floatNames []string,
-) {
+func (fw *flowsProcessor) report(w *worker, ts time.Time, flow *biFlow, isOver bool, intNames, uintNames, floatNames []string) {
 	event := createEvent(fw.watcher, ts, flow, isOver, intNames, uintNames, floatNames)
 
 	debugf("add event: %v", event)
 	fw.spool.publish(event)
 }
 
-func createEvent(
-	watcher procs.ProcessesWatcher,
-	ts time.Time, f *biFlow,
-	isOver bool,
-	intNames, uintNames, floatNames []string,
-) beat.Event {
+func createEvent(watcher procs.ProcessesWatcher, ts time.Time, f *biFlow, isOver bool, intNames, uintNames, floatNames []string) beat.Event {
 	timestamp := ts
 
 	event := mapstr.M{
@@ -442,10 +573,7 @@ func createEvent(
 	}
 }
 
-func encodeStats(
-	stats *flowStats,
-	ints, uints, floats []string,
-) map[string]interface{} {
+func encodeStats(stats *flowStats, ints, uints, floats []string) map[string]interface{} {
 	report := make(map[string]interface{})
 
 	i := 0
@@ -519,4 +647,43 @@ func putOrAppendUint64(m mapstr.M, key string, value uint64) {
 			m[key] = append(v, value)
 		}
 	}
+}
+
+// spool is an event publisher spool.
+type spool struct {
+	pub    Reporter
+	events []beat.Event
+}
+
+// init sets the destination and spool size.
+func (s *spool) init(pub Reporter, sz int) {
+	s.pub = pub
+	s.events = make([]beat.Event, 0, sz)
+}
+
+// publish queues the event for publication, flushing to the destination
+// if the spool is full.
+func (s *spool) publish(event beat.Event) {
+	s.events = append(s.events, event)
+	if len(s.events) == cap(s.events) {
+		s.flush()
+	}
+}
+
+// flush sends the spooled events to the destination and clears them
+// from the spool.
+func (s *spool) flush() {
+	if len(s.events) == 0 {
+		return
+	}
+	s.pub(s.events)
+	// A newly allocated spool is created since the
+	// elements of s.events are no longer owned by s
+	// during testing and mutating them causes a panic.
+	//
+	// The beat.Client interface which Reporter is
+	// derived from is silent on whether the caller
+	// is allowed to modify elements of the slice
+	// after the call to the PublishAll method returns.
+	s.events = make([]beat.Event, 0, cap(s.events))
 }
